+++ src/cluster.tsx
@@ -4,8 +4,8 @@ import clsx from 'clsx'
 import type {DashTokens} from '@dash-ui/styles'
 import {Box} from './box'
 import type {BoxProps} from './box'
-import {useLayout} from './layout'
-import type {ResponsiveProp, ResponsivePropCallback} from './layout'
+import {useResponsiveStyles} from './layout'
+import type {ResponsiveProp, ResponsiveLazyProp} from './layout'
 import {justifyContent, flexDirection} from './styles'
 import {forwardRefAs} from './utils'
 
@@ -40,7 +40,7 @@ export const Cluster = forwardRefAs<ClusterProps, 'div'>(function Cluster(
   {className, gap, reverse = false, ...props},
   ref
 ) {
-  const {responsiveStyles, styles} = useLayout()
+  const styles = useResponsiveStyles()
 
   return (
     <Box
@@ -55,8 +55,8 @@ export const Cluster = forwardRefAs<ClusterProps, 'div'>(function Cluster(
               flex-shrink: 0;
             }
           `,
-          responsiveStyles(gapStyle(reverse)).css(gap),
-          responsiveStyles(reverseStyle).css(reverse)
+          gap === void 0 ? '' : styles.lazy(gapStyle(reverse)).css(gap),
+          reverse === void 0 ? '' : styles.lazy(reverseStyle).css(reverse)
         )
       )}
       {...props}
@@ -70,7 +70,10 @@ const reverseStyle = (reverse: boolean) =>
 const gapStyle = (
   reverse: ClusterProps['reverse']
   // @ts-expect-error
-): ResponsivePropCallback<keyof DashTokens['gap']> => (gapProp, queryName) => {
+): ResponsiveLazyProp<Extract<keyof DashTokens['gap'], string | number>> => (
+  gapProp,
+  queryName
+) => {
   const reversed =
     !reverse || typeof reverse === 'boolean' ? reverse : reverse[queryName]
   const marginDirection = reversed ? 'right' : 'left'
@@ -96,8 +99,10 @@ export interface ClusterProps extends Omit<BoxProps, 'display'> {
    * Sets a vertical and horizontal gap between the child elements in the
    * cluster using the "gap" token in your theme
    */
-  // @ts-expect-error
-  readonly gap?: ResponsiveProp<keyof DashTokens['gap']>
+  readonly gap?: ResponsiveProp<
+    // @ts-expect-error
+    Extract<keyof DashTokens['gap'], string | number>
+  >
 }
 
 /* istanbul ignore next */